requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
chromadb
langchain-openai
sentence-transformers
langchain_huggingface
faiss-cpu
langchain_chroma
duckdb
pandas
openai
streamlit